Evaluate language arguments and pass as 'character'
Previously, arguments in designers with class 'call' were not being used as inputs in the designer function.

This commit evaluates these language inputs (e.g., `rep(0, 2^k)` in `outcome_means` parameter of `factorial_designer()`) and turns them into a string input of the form "0, 0, 0, 0" based on the default value of `k`.